Added support for TSpace in the fragmenter.  Also added a 'big' option to parsing to handle remaining problem segments after the normal parsing pass.

+++ src/main/java/edu/uiowa/slis/incite/Fragmenter.java
@@ -11,6 +11,9 @@ import java.util.Vector;
 import org.apache.log4j.Logger;
 import org.apache.log4j.PropertyConfigurator;
 
+import com.ibm.tspaces.TupleSpace;
+import com.ibm.tspaces.TupleSpaceException;
+
 import edu.uiowa.lex.basicLexerToken;
 import edu.uiowa.lex.lexer;
 import edu.uiowa.pos_tagging.POStagger;
@@ -21,6 +24,9 @@ public class Fragmenter  {
     static Logger logger = Logger.getLogger(Fragmenter.class);
     static Connection conn = null;
     static LocalProperties prop_file = PropertyLoader.loadProperties("incite");
+    static boolean useTSpace = true;
+    static TupleSpace ts = null;
+    static final String host = "deep-thought.slis.uiowa.edu";
 
     int count = 0;
     lexer theLexer = null;
@@ -30,8 +36,17 @@ public class Fragmenter  {
     public static void main(String[] args) throws Exception {
 	PropertyConfigurator.configure(args[0]);
 
+	if (useTSpace) {
+	    logger.info("initializing tspace...");
+	    try {
+		ts = new TupleSpace("incite", host);
+	    } catch (TupleSpaceException tse) {
+		logger.error("TSpace error: " + tse);
+	    }
+	}
+	
 	conn = getConnection();
-	new Fragmenter();
+	new Fragmenter(args);
     }
     
     static void loadConfiguration() {
@@ -51,30 +66,40 @@ public class Fragmenter  {
 	return conn;
     }
 
-    public Fragmenter() throws Exception {
-	threadedGenerate();
+    public Fragmenter(String[] args) throws Exception {
+	threadedGenerate(args);
     }
 
-    void threadedGenerate() throws Exception {
+    void threadedGenerate(String[] args) throws Exception {
 	int maxCrawlerThreads = Runtime.getRuntime().availableProcessors();
 	Vector<Integer> queue = new Vector<Integer>();
 	Thread[] fragmenterThreads = new Thread[maxCrawlerThreads];
-	
-	logger.info("truncating fragment...");
-	conn.prepareStatement("truncate " + prop_file.getProperty("jdbc.schema") + ".fragment").execute();
-	conn.commit();
 
-	logger.info("loading queue...");
+	if (!useTSpace || (useTSpace && args.length > 1 && args[1].equals("-hub"))) {
+	    logger.info("truncating fragment...");
+	    conn.prepareStatement("truncate " + prop_file.getProperty("jdbc.schema") + ".fragment").execute();
+	    conn.commit();
+
+	    logger.info("loading queue...");
 	    PreparedStatement stmt = conn.prepareStatement("select distinct id from extraction.parse where not exists (select * from extraction.fragment where fragment.id=parse.id) order by id");
 	    ResultSet rs = stmt.executeQuery();
 	    while (rs.next()) {
-		int ID = rs.getInt(1);
-		queue.add(ID);
+		if (useTSpace) {
+		    try {
+			ts.write("fragment_request", rs.getInt(1));
+		    } catch (Exception e) {
+			logger.error("tspace exception raised: ", e);
+		    }
+		} else
+		    queue.add(rs.getInt(1));
+	    }
+	    logger.info("queue loaded.");
+	    if (useTSpace)
+		return;
 	}
-	logger.info("queue loaded.");
 
 	for (int i = 0; i < maxCrawlerThreads; i++) {
-	    Thread theThread = new Thread(new FragmenterThread(i, queue, getConnection(), prop_file));
+	    Thread theThread = new Thread(new FragmenterThread(i, queue, getConnection(), prop_file, useTSpace));
 	    theThread.setPriority(Math.max(theThread.getPriority() - 2, Thread.MIN_PRIORITY));
 	    theThread.start();
 	    fragmenterThreads[i] = theThread;
@@ -83,10 +108,12 @@ public class Fragmenter  {
 	    fragmenterThreads[i].join();
 	}
 
-	logger.info("refreshing fragments...");
-	conn.prepareStatement("refresh materialized view " + prop_file.getProperty("jdbc.schema") + ".fragments").execute();
-	conn.commit();
-	logger.info("done.");
+	if (!useTSpace || (useTSpace && args.length > 1 && args[1].equals("-hub"))) {
+	    logger.info("refreshing fragments...");
+	    conn.prepareStatement("refresh materialized view " + prop_file.getProperty("jdbc.schema") + ".fragments").execute();
+	    conn.commit();
+	    logger.info("done.");
+	}
     }
 
 }
